% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.LASERC.R
\name{cv.LASERC}
\alias{cv.LASERC}
\title{Cross-validation for selecting L in LASERC}
\usage{
cv.LASERC(
  dat,
  batch,
  mod,
  standardize = TRUE,
  center = NULL,
  scale = NULL,
  ...,
  L_set,
  K = 3,
  parallel = TRUE,
  ncores = 3,
  seed = 1234
)
}
\arguments{
\item{dat}{Numeric matrix with subjects as rows,
and functional connectivity features as columns.}

\item{batch}{Numeric or character vector specifying the batch/site
variable needed for harmonization.}

\item{mod}{Optional model matrix for outcome of interest and
other covariates besides batch/scanner.}

\item{standardize}{Logical indicating whether to standardize the columns of \code{dat} prior to
model fitting. Defaults to \code{TRUE}.}

\item{center}{A numeric vector of length \code{p} giving column means for centering (if \code{standardize=TRUE}).
If \code{NULL}, column means of \code{dat} are used.}

\item{scale}{A numeric vector of length \code{p} giving column standard deviations for scaling (if \code{standardize=TRUE}).
If \code{NULL}, column standard deviations of \code{dat} are used.}

\item{\dots}{Additional parameters passed to \code{\link{LASERC}}, such as \code{lambda},
\code{tau}, or \code{MaxIteration}.}

\item{L_set}{A numeric vector of candidate \code{L} to be evaluated via cross-validation.}

\item{K}{An integer specifying the number of folds for cross-validation. Default is 3.}

\item{parallel}{Logical indicating whether to parallelize the cross-validation loop
using \pkg{doParallel} and \pkg{foreach}. Defaults to \code{TRUE}.}

\item{ncores}{An integer specifying the number of CPU cores to register when
\code{parallel = TRUE}. Default is 3.}

\item{seed}{An integer seed used for reproducibility in creating folds. Default is 1234.}
}
\value{
An object of class \code{"cv.LASERC"}, which is a list containing:
\describe{
  \item{\code{L_set}}{The vector of \code{L} values tested.}
  \item{\code{cv.mean}}{A numeric vector of mean loss across folds for each \code{L}.}
  \item{\code{cv.se}}{A numeric vector of standard deviations of the loss across folds
  for each \code{L}.}
  \item{\code{L_est}}{The chosen \code{L} (the value that minimizes the cross-validation loss).}
}
}
\description{
\code{cv.LASERC} performs k-fold cross-validation over a grid of possible
ranks (\code{L_set}) for the \code{LASERC} function.
}
\examples{
\dontrun{
data(sim_data)
dat=sim_data$dat
batch=sim_data$batch
mod=sim_data$mod
# Candidate L values
L_candidates <- 1:10

# Perform 3-fold CV for each candidate L
cv_results <- cv.LASERC(
  dat = dat,
  batch = batch,
  mod = mod,
  L_set = L_candidates,
  K = 3,
  parallel = FALSE  # set to TRUE to test parallel mode
)
plot(cv_results)
}
}
