% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASERC.R
\name{LASERC}
\alias{LASERC}
\title{LASERC harmonization}
\usage{
LASERC(
  dat,
  batch,
  mod = NULL,
  L = 3,
  standardize = TRUE,
  center = NULL,
  scale = NULL,
  MaxIteration = 100,
  lambda = log(n * V)/(n * V),
  tau = 0.5 * sqrt(1/(V)),
  espli = 1e-04,
  silent = FALSE,
  initial = "eigen",
  Theta_ini = NULL,
  penalty
)
}
\arguments{
\item{dat}{Numeric matrix with subjects as rows \code{(n)},
and functional connectivity features as columns \code{(p)}.}

\item{batch}{Numeric or character vector specifying the batch/site
variable needed for harmonization.}

\item{mod}{Optional model matrix for outcome of interest and
other covariates besides batch/scanner.}

\item{L}{Numeric specifying the number of sparsity connectivity patterns that needs to be tuned (default = 3).}

\item{standardize}{Logical indicating whether to standardize the columns of \code{dat} prior to
model fitting. Defaults to \code{TRUE}.}

\item{center}{A numeric vector of length \code{p} giving column means for centering (if \code{standardize=TRUE}).
If \code{NULL}, column means of \code{dat} are used.}

\item{scale}{A numeric vector of length \code{p} giving column standard deviations for scaling (if \code{standardize=TRUE}).
If \code{NULL}, column standard deviations of \code{dat} are used.}

\item{MaxIteration}{An integer specifying the maximum number of iterations for each stage of the
iterative procedure. Default is 200.}

\item{lambda}{Numeric scalar controlling the TLP penalty (default = \code{log(n)/n}). Used as the constant part in TLP.}

\item{tau}{Numeric scalar controlling a threshold parameter in TLP (default = \code{0.3 * sqrt(1/(0.5*V))}).}

\item{espli}{A numeric convergence threshold for the relative change in the loss function across
iterations (default = 1e-4).}

\item{silent}{Logical indicating whether to suppress iteration progress messages. Defaults to \code{FALSE}.}

\item{initial}{A character string specifying the initial estimation method for
\code{(A, S, U)}. Possible values are \code{"eigen"} or \code{"prespecify"}.
Defaults to \code{"eigen"}.}

\item{Theta_ini}{A list containing user-specified initial values for \code{A}, \code{S},
\code{U}, \code{Fmat}, \code{sigma2}, and \code{phi2} if \code{initial = "prespecify"}.
Ignored otherwise.}
}
\value{
A list with the following components:
\describe{
  \item{\code{dat.LASERC}}{The harmonized (corrected) data matrix after adjusting for batch
  effects.}
  \item{\code{estimates}}{A list of final parameter estimates, including \code{A}, \code{S},
  \code{U}, \code{xFmat}, \code{sigma2}, \code{phi2}, \code{Fmat}, \code{gamma}, \code{beta},
  and \code{alpha}.}
  \item{\code{dat.original}}{The original data matrix (n x p) passed to the function.}
  \item{\code{batch}}{The batch vector used for correction.}
}
}
\description{
Main function to perform LASERC harmonization.
}
\examples{
\dontrun{
data(sim_data)
dat=sim_data$dat
batch=sim_data$batch
mod=sim_data$mod
n=nrow(dat)
p=ncol(dat)
V=(sqrt(1+8*p)-1)/2
results <- LASERC(
  dat = dat,
  batch = batch,
  mod = mod,
  L = 5
)
}
}
